Improvements to playlist state saving.
Introduced a temp file in case of an I/O error (such as a full disk),
and trigger save during deferred stop and ondestroy to fix a pragmatic
issue where the service might be killed while waiting for deferred stop.

An ideal solution would be to implement playlist state saving in onStop
which I believe is called even when the low memory killer is
responsible.  More research to follow.

git-svn-id: http://five.googlecode.com/svn/trunk/android-client@1123 0f5e3bc8-7f4c-0410-bf3f-39b466953847