Fixed a critical bug which could corrupt downloaded file.
If an exception was thrown during connect or parsing of the headers and
when resuming a partially cached file, the subsequent retry attempt
would start the file downloading at the beginning but append the result
to the partial cache file.

Also revised Content-Range parser to actually comply with the standard
which brings compatability with the new five server.

git-svn-id: http://five.googlecode.com/svn/trunk/android-client@1128 0f5e3bc8-7f4c-0410-bf3f-39b466953847