Significant performance optimizations to sync path.
We now enter a transaction for both the population of the temp database
as well as the main.  In order to accomplish this we must access a new
API from Android 2.0 which gives us access to the local content provider
instance (pre-2.0 is supported through reflection into a hidden API).

Indices are removed from the temp database to slightly improve insert
performance.

Additionally, crude timing code has been introduced which is used to
debug wall time during various stages of the sync process.

git-svn-id: http://five.googlecode.com/svn/trunk/android-client@1132 0f5e3bc8-7f4c-0410-bf3f-39b466953847