New sync UI.
Removed "Five Sync" launcher icon, merged with the "Five Sync" main
screen.  Replaced the "Songs" main button with a "Settings" button which
is the central area for adding/editing sources as well as observing the
sync progress.

Removed a significant amount of code and simplified numerous areas of
the UI and service design.

Additional work left to be done includes automatically prompting to add
a server when there are no sources, implementing auto-sync, and adding a
notification during sync which links back to the settings screen.

git-svn-id: http://five.googlecode.com/svn/trunk/android-client@1139 0f5e3bc8-7f4c-0410-bf3f-39b466953847