Fixed an unlikely race condition which could corrupt sync.
After each sync, we previously record the latest of all synced items (in
each of the 5 synced data sets) and use that as the lastModifiedSince
argument upon the next sync.  This creates an issue if an item is
modified from one data set while another is syncing.

This change computes the lastModifiedSince argument on each sync by
examining the latest sync time of each data set.

Additionally, the last sync time that's displayed to the user has been
changed to accurately reflect the timestamp of sync completion.

git-svn-id: http://five.googlecode.com/svn/trunk/android-client@1149 0f5e3bc8-7f4c-0410-bf3f-39b466953847