Orientation change on Settings main page erroneously triggers SourceAdd.
The Intent we deliver to Settings remains in the OOBE state after the
first source is added and the initial sync occurs.  On this screen, an
orientation change would then bring the user back to the add source
screen.  Fixed this issue by simply checking if any sources exist before
calling into SourceAdd.