Refactored Player screen.
Most of the UI has been moved into a new composite control
PlayerControls.  This largely replaces and extends what
BetterReflectionLayout was doing before, but using a more
straight-forward design.

The basic proportion of the album artwork to the main controls panel
(prev/pause/next) has been adjusted so that it will look correct on
screens with different aspect ratios (such as the Nexus One).