Significantly refactored Artist/AlbumList.
These two classes are now utilizing a much more modern approach to
Android activity design and are factored such that they can share a
greater amount of code between them.

Additionally, the way images are loaded has been reworked to use a
design somewhat similar to the contacts app.  This design is simpler and
doesn't rely on a forked version of FastScrollView.